Move magic number inspection to disabled
