+++ plugin/src/main/kotlin/org/jetbrains/research/pynose/plugin/inspections/pytest/disabled/MagicNumberTestTestSmellPytestInspection.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.pynose.plugin.inspections.pytest
+package org.jetbrains.research.pynose.plugin.inspections.pytest.disabled
 
 import com.intellij.codeInspection.LocalInspectionToolSession
 import com.intellij.codeInspection.ProblemsHolder
