+++ plugin/src/test/kotlin/org/jetbrains/research/pynose/plugin/inspections/pytest/disabled/MagicNumberTestTestSmellPytestInspectionTests.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.pynose.plugin.inspections.unittest
+package org.jetbrains.research.pynose.plugin.inspections.pytest.disabled
 
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.components.service
@@ -11,7 +11,7 @@ import org.jetbrains.research.pynose.plugin.util.TestSmellBundle
 import org.junit.Test
 import org.junit.jupiter.api.BeforeAll
 
-class MagicNumberTestTestSmellUnittestInspectionTests : AbstractTestSmellInspectionTestWithSdk() {
+class MagicNumberTestTestSmellPytestInspectionTests: AbstractTestSmellInspectionTestWithSdk() {
 
     @BeforeAll
     override fun setUp() {
@@ -19,49 +19,42 @@ class MagicNumberTestTestSmellUnittestInspectionTests : AbstractTestSmellInspect
         mockkObject(myFixture.project.service<TestRunnerServiceFacade>())
         every {
             myFixture.project.service<TestRunnerServiceFacade>().getConfiguredTestRunner(any())
-        } returns TestRunner.UNITTESTS
-        myFixture.enableInspections(MagicNumberTestTestSmellUnittestInspection())
+        } returns TestRunner.PYTEST
+        myFixture.enableInspections(MagicNumberTestTestSmellPytestInspection())
     }
 
     override fun getTestDataPath(): String {
-        return "src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/magic_number/unittest"
+        return "src/test/resources/org/jetbrains/research/pynose/plugin/inspections/data/magic_number/pytest"
     }
 
     @Test
     fun `test highlighted magic number`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
+            "test_file.py", "class TestClass:\n" +
                     "    def test_something(self):\n" +
-                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.magic.number.description")}\">assert 17 == 1</weak_warning>\n" +
+                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.magic.number.description")}\">assert 1 == 41</weak_warning>\n" +
                     "        assert \"H\" != \"F\"\n" +
-                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.magic.number.description")}\">self.assertTrue(22 == 1 + 2)</weak_warning>\n"
+                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.magic.number.description")}\">assert 24 == 1 + 2</weak_warning>\n"
         )
-
         myFixture.checkHighlighting()
     }
 
     @Test
     fun `test highlighted combination of numbers and letters`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass(unittest.TestCase):\n" +
-                    "    def test_something(self):\n" +
-                    "        <weak_warning descr=\"${TestSmellBundle.message("inspections.magic.number.description")}\">self.assertEqual(\"G\", 882)</weak_warning>"
+            "test_file.py", "def test_something(self):\n" +
+                    "    <weak_warning descr=\"${TestSmellBundle.message("inspections.magic.number.description")}\">assert \"G\" != 27</weak_warning>"
         )
-
         myFixture.checkHighlighting()
     }
 
     @Test
-    fun `test magic number without unittest dependency`() {
+    fun `test magic number wrong class name`() {
         myFixture.configureByText(
-            "test_file.py", "import unittest\n" +
-                    "class SomeClass():\n" +
+            "test_file.py", "class SomeClass():\n" +
                     "    def test_something(self):\n" +
-                    "        assert 1 == 81"
+                    "        assert 1 == 51"
         )
-
         val highlightInfos = myFixture.doHighlighting()
         assertTrue(!highlightInfos.any { it.severity == HighlightSeverity.WEAK_WARNING })
     }
@@ -69,7 +62,6 @@ class MagicNumberTestTestSmellUnittestInspectionTests : AbstractTestSmellInspect
     @Test
     fun `test magic number multiple`() {
         myFixture.configureByFile("test_magic_number_multiple.py")
-
         myFixture.checkHighlighting()
     }
 }
